rust: move key::Key impl to key::callback
